Rewrite much of the GM ticket system * Extract storage and manipulation of tickets to TicketMgr (from ObjectMgr) * Extract ticket commands to TicketCommands.cpp * Adds support for sending GM responses and GM surveys. * Fix structure of several ticket-related packets. * Add support for understanding lag reports. * Thanks Zor for some of the packet structures, and Cyrax for some sniffs * Please report any issues encountered via the tracker.
--HG--
branch : trunk

+++ src/server/authserver/Server/AuthSocket.cpp
@@ -907,7 +907,7 @@ bool AuthSocket::_HandleXferResume()
     uint64 start;
     socket().recv_skip(1);
     socket().recv((char*)&start,sizeof(start));
-    fseek(pPatch, start, 0);
+    fseek(pPatch, long(start), 0);
 
     ACE_Based::Thread u(new PatcherRunnable(this));
     return true;
